BE: (#0004) renamed database package to db
